IN-196 fix: fix open dialog
